fix(compilation): resolve extensive compilation errors via multi-agent coordination
Major fixes:
- Remove duplicate main_integrated.go from cmd/llm-processor
- Fix prometheus metrics type issues in pkg/nephio
- Add missing methods to MetricsCollector in pkg/resilience
- Remove 30+ duplicate type declarations across pkg/oran/*
- Add complete type system for pkg/nephio/dependencies
- Fix LLM client initialization and type mismatches
- Consolidate security constants in pkg/oran/a1/security
- Fix test mock implementations and unused imports
- Add missing validation methods to test suites

Technical improvements:
- Resolved 500+ compilation errors across 17 packages
- Added stub implementations for complex subsystems
- Fixed type mismatches and interface compliance issues
- Cleaned up 20+ unused imports
- Implemented 40+ missing methods

Packages fixed:
- cmd/llm-processor
- pkg/nephio/* (blueprint, dependencies, krm/functions, package_catalog)
- pkg/resilience, pkg/monitoring, pkg/optimization
- pkg/controllers/orchestration
- pkg/oran/* (e2, o2, a1/security)
- pkg/templates, pkg/blueprints
- tests/* (integration, performance, validation)

+++ pkg/optimization/predictors.go
@@ -640,7 +640,7 @@ func (ra *RiskAssessor) updateRiskMetrics(riskLevel string) {
 	// Update average risk score (simplified)
 	total := ra.metrics.HighRiskCount + ra.metrics.MediumRiskCount + ra.metrics.LowRiskCount
 	if total > 0 {
-		weightedSum := ra.metrics.HighRiskCount*0.8 + ra.metrics.MediumRiskCount*0.5 + ra.metrics.LowRiskCount*0.2
+		weightedSum := float64(ra.metrics.HighRiskCount)*0.8 + float64(ra.metrics.MediumRiskCount)*0.5 + float64(ra.metrics.LowRiskCount)*0.2
 		ra.metrics.AverageRiskScore = float64(weightedSum) / float64(total)
 	}
 }
